Refactor code to keep most drop range analysis in drop_ranges.rs
